DomainTip: move component to blocks
